Move java classes into brave/browser for quick_search_engines
