+++ android/brave_java_sources.gni
@@ -311,15 +311,6 @@ brave_java_sources = [
   "../../brave/android/java/org/chromium/chrome/browser/qrreader/BarcodeTrackerFactory.java",
   "../../brave/android/java/org/chromium/chrome/browser/qrreader/CameraSource.java",
   "../../brave/android/java/org/chromium/chrome/browser/qrreader/CameraSourcePreview.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/ItemTouchHelperCallback.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/settings/QuickSearchEnginesAdapter.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/settings/QuickSearchEnginesCallback.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/settings/QuickSearchEnginesFragment.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/settings/QuickSearchEnginesModel.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/settings/QuickSearchEnginesSettingsViewHolder.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/utils/QuickSearchEnginesUtil.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/views/QuickSearchEnginesViewAdapter.java",
-  "../../brave/android/java/org/chromium/chrome/browser/quick_search_engines/views/QuickSearchEnginesViewHolder.java",
   "../../brave/android/java/org/chromium/chrome/browser/quickactionsearchwidget/BraveQuickActionSearchWidgetProvider.java",
   "../../brave/android/java/org/chromium/chrome/browser/rate/BraveAskPlayStoreRatingDialog.java",
   "../../brave/android/java/org/chromium/chrome/browser/rate/BraveRateDialogFragment.java",
@@ -459,7 +450,6 @@ brave_java_sources = [
   "../../brave/android/java/org/chromium/chrome/browser/util/KeyboardVisibilityHelper.java",
   "../../brave/android/java/org/chromium/chrome/browser/util/LiveDataUtil.java",
   "../../brave/android/java/org/chromium/chrome/browser/util/PackageUtils.java",
-  "../../brave/android/java/org/chromium/chrome/browser/util/SharedPreferencesHelper.java",
   "../../brave/android/java/org/chromium/chrome/browser/util/TabUtils.java",
   "../../brave/android/java/org/chromium/chrome/browser/util/TouchDelegateComposite.java",
   "../../brave/android/java/org/chromium/chrome/browser/util/Triple.java",
