+++ android/java/org/chromium/chrome/browser/app/BraveActivity.java
@@ -157,11 +157,11 @@ import org.chromium.chrome.browser.prefetch.settings.PreloadPagesState;
 import org.chromium.chrome.browser.privacy.settings.BravePrivacySettings;
 import org.chromium.chrome.browser.profiles.Profile;
 import org.chromium.chrome.browser.profiles.ProfileManager;
-import org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesCallback;
-import org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesFragment;
-import org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesModel;
-import org.chromium.chrome.browser.quick_search_engines.utils.QuickSearchEnginesUtil;
-import org.chromium.chrome.browser.quick_search_engines.views.QuickSearchEnginesViewAdapter;
+import org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesCallback;
+import org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesFragment;
+import org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesModel;
+import org.chromium.brave.browser.quick_search.utils.QuickSearchEnginesUtil;
+import org.chromium.brave.browser.quick_search.views.QuickSearchEnginesViewAdapter;
 import org.chromium.chrome.browser.rate.BraveRateDialogFragment;
 import org.chromium.chrome.browser.rate.RateUtils;
 import org.chromium.chrome.browser.rewards.adaptive_captcha.AdaptiveCaptchaHelper;
