+++ android/java/res/xml/brave_search_engines_preferences.xml
@@ -17,7 +17,7 @@
         android:order="2"
         android:title="@string/prefs_private_search_engine"/>
     <org.chromium.components.browser_ui.settings.ChromeBasePreference
-        android:fragment="org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesFragment"
+        android:fragment="org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesFragment"
         android:key="quick_search_engine"
         android:order="3"
         android:title="@string/quick_search_engines_with_hyphen"/>
