+++ browser/search_engines/android/java/src/org/chromium/brave/browser/SharedPreferencesHelper.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.util;
+package org.chromium.brave.browser;
 
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
