+++ browser/search_engines/android/java/src/org/chromium/brave/browser/quick_search/views/QuickSearchEnginesViewAdapter.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.quick_search_engines.views;
+package org.chromium.brave.browser.quick_search.views;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -13,10 +13,10 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import org.chromium.chrome.R;
-import org.chromium.chrome.browser.app.BraveActivity;
-import org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesCallback;
-import org.chromium.chrome.browser.quick_search_engines.settings.QuickSearchEnginesModel;
+import org.chromium.brave.browser.quick_search.R;
+import org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesCallback;
+import org.chromium.brave.browser.quick_search.settings.QuickSearchEnginesModel;
+import org.chromium.brave.browser.quick_search.utils.QuickSearchEnginesUtil;
 
 import java.util.List;
 
@@ -44,11 +44,11 @@ public class QuickSearchEnginesViewAdapter
                 && quickSearchEnginesModel.getType()
                         == QuickSearchEnginesModel.QuickSearchEnginesModelType.AI_ASSISTANT) {
             quickSearchViewHolder.mSearchEngineLogo.setImageResource(R.drawable.ic_leo_icon);
-        } else if (BraveActivity.YOUTUBE_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
+        } else if (QuickSearchEnginesUtil.YOUTUBE_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
             quickSearchViewHolder.mSearchEngineLogo.setImageResource(R.drawable.ic_social_youtube);
-        } else if (BraveActivity.BING_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
+        } else if (QuickSearchEnginesUtil.BING_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
             quickSearchViewHolder.mSearchEngineLogo.setImageResource(R.drawable.ic_microsoft_color);
-        } else if (BraveActivity.STARTPAGE_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
+        } else if (QuickSearchEnginesUtil.STARTPAGE_SEARCH_ENGINE_KEYWORD.equals(keyword)) {
             quickSearchViewHolder.mSearchEngineLogo.setImageResource(R.drawable.ic_startpage_color);
         } else {
             mQuickSearchEnginesCallback.loadSearchEngineLogo(
