+++ browser/search_engines/android/java/src/org/chromium/brave/browser/quick_search/views/QuickSearchEnginesViewHolder.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.quick_search_engines.views;
+package org.chromium.brave.browser.quick_search.views;
 
 import android.view.View;
 import android.widget.ImageView;
